# save-backend/backend-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SAVE Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5800",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/projects/get-for-current-user": {
      "get": {
        "description": "Get list of projects related to current user",
        "operationId": "getProjectsOfCurrentUser",
        "parameters": [
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Successfully fetched users from project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get projects of current authenticated user",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{organizationName}/{projectName}/users": {
      "get": {
        "description": "Get list of users that are connected with given project and their roles in it.",
        "operationId": "getAllUsersByProjectNameAndOrganizationName",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "description": "Successfully fetched users from project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "description": "Project with such name was not found."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get users from project with their roles.",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{organizationName}/{projectName}/users/not-from": {
      "get": {
        "description": "Get list of users that are not connected with given project.",
        "operationId": "getAllUsersNotFromProjectWithNamesStartingWith",
        "parameters": [
          {
            "description": "name of an organization",
            "in": "path",
            "name": "organizationName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "name of a project",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "example": "basic",
            "in": "header",
            "name": "X-Authorization-Source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "description": "Successfully fetched users not from project."
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "description": "Project with such name was not found or considered to be private."
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "Get users not from project.",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Organization": {
        "required": [
          "canCreateContests",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string"
          },
          "canCreateContests": {
            "type": "boolean"
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "DELETED"
            ]
          }
        }
      },
...
      "Project": {
        "required": [
          "contestRating",
          "name",
          "numberOfContainers",
          "organization",
          "public",
          "status"
        ],
        "type": "object",
        "properties": {
          "contestRating": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "numberOfContainers": {
            "type": "integer",
            "format": "int32"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "public": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "DELETED"
            ]
          },
          "url": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UserInfo": {
        "required": [
          "isActive",
          "name",
          "organizations",
          "originalLogins",
          "projects"
        ],
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "gitHub": {
            "type": "string"
          },
          "globalRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "NONE",
              "OWNER",
              "SUPER_ADMIN",
              "VIEWER"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "isActive": {
            "type": "boolean"
          },
          "linkedin": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "oldName": {
            "type": "string"
          },
          "organizations": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ADMIN",
                "NONE",
                "OWNER",
                "SUPER_ADMIN",
                "VIEWER"
              ]
            }
          },
          "originalLogins": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projects": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ADMIN",
                "NONE",
                "OWNER",
                "SUPER_ADMIN",
                "VIEWER"
              ]
            }
          },
          "source": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  }
}